Add incomplete RedBlackTree
No more work is to be done on this branch anymore. In fact, no more work
is to be done to maintain the C data structures portion of the code.
After working strictly in C, the code has become unmaintainable.

The next branch of this project will abandon its goal to remain C89
compatible and begin the inclusion of C++98 in areas that would benefit
from it.