Remove all Windows VS_FIXEDFILEINFO code
It turns out that the wrapper functions is highly inefficient for
performing multiple calls to VerQueryValue. It would be faster to simply
store the result of GetFileVersionInfo and reuse it. This would avoid
unwanted malloc and free calls on kilobytes of data. Doing this would
effectively render a wrapper function moot.

C++ perhaps may be more suited for this task.